Fix unit test

+++ src/test/EquationTest.java
@@ -13,14 +13,14 @@ class EquationTest {
 	@Test
 	void createEquation() {
 		BigDecimal three = new BigDecimal(3);
-		assertEquals(13, new Equation("k + 5 * 2", "k").evaluate(three));
-		assertEquals(17, new Equation("k * 5 + 2", "k").evaluate(three));
-		assertEquals(13.5, new Equation("k * 5 - 2 + 1 / 2", "k").evaluate(three));
-		assertEquals(16.5, new Equation("k * 5 + 2 - 1 / 2", "k").evaluate(three));
-		assertEquals(-14, new Equation("k - 2 - 3*5", "k").evaluate(three));
-		assertEquals(8, new Equation("k - (2 - 3) * 5", "k").evaluate(three));
-		assertEquals(20, new Equation("(k - (2 - 3)) * 5", "k").evaluate(three));
-		assertEquals(16, new Equation("k - (2 - (3 * 5))", "k").evaluate(three));
+		assertEquals(13, ((BigDecimal) new Equation("k + 5 * 2", "k").evaluate(three)).intValue());
+		assertEquals(17, ((BigDecimal) new Equation("k * 5 + 2", "k").evaluate(three)).intValue());
+		assertEquals(13.5, ((BigDecimal) new Equation("k * 5 - 2 + 1 / 2", "k").evaluate(three)).doubleValue());
+		assertEquals(16.5, ((BigDecimal) new Equation("k * 5 + 2 - 1 / 2", "k").evaluate(three)).doubleValue());
+		assertEquals(-14, ((BigDecimal) new Equation("k - 2 - 3*5", "k").evaluate(three)).intValue());
+		assertEquals(8, ((BigDecimal) new Equation("k - (2 - 3) * 5", "k").evaluate(three)).intValue());
+		assertEquals(20, ((BigDecimal) new Equation("(k - (2 - 3)) * 5", "k").evaluate(three)).intValue());
+		assertEquals(16, ((BigDecimal) new Equation("k - (2 - (3 * 5))", "k").evaluate(three)).intValue());
 	}
 
 }
